Roles y permisos con Laravel Permission

+++ app/Http/Controllers/UsuariosController.php
@@ -5,19 +5,25 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreUsuarioRequest;
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 
 class UsuariosController extends Controller
 {
     public function index() {
+        Gate::authorize('ver usuarios');
+
         $usuarios = User::paginate(5);
         return view('usuarios.index', compact('usuarios'));
     }
 
     public function create() {
+        Gate::authorize('crear usuarios');
+
         return view('usuarios.create');
     }
 
     public function store(StoreUsuarioRequest $request) {
+        Gate::authorize('crear usuarios');
         // 1. Validar los datos (en StoreUsuarioRequest)
         // 1.5) Procesar y guardar la imagen de perfil
         /*if($request->hasFile('avatar')) {
@@ -37,6 +43,8 @@ class UsuariosController extends Controller
     }
 
     public function show(int $id) {
+        Gate::authorize('ver usuarios');
+
         $usuario = User::with('listasLectura.libros')->findOrFail($id);
 
         $todosLosLibros = $usuario->listasLectura->flatMap(function ($lista) {
@@ -51,12 +59,16 @@ class UsuariosController extends Controller
     }
 
     public function edit(int $id) {
+        Gate::authorize('editar usuarios');
+
         $usuario = User::findOrFail($id);
 
         return view('usuarios.edit', compact('usuario'));
     }
 
     public function update(StoreUsuarioRequest $request, int $id) {
+        Gate::authorize('editar usuarios');
+
         // Actualizar en la BD
         User::where('id', $id)
             ->update([
@@ -70,6 +82,8 @@ class UsuariosController extends Controller
     }
 
     public function destroy(int $id) {
+        Gate::authorize('eliminar usuarios');
+
         // Eliminar de la BD
         User::destroy($id);
